Make <selectmenu>'s parts search always a flat tree traversal that skips nested <selectmenu>s.
Per an OpenUI resolution [1], a <selectmenu>'s search for parts should
be a flat tree traversal that doesn't descend into nested <selectmenu>
or <select> elements.

This change updates <selectmenu> to use the resolved-upon behavior.
As a consequence, the parts chosen depend only on tree structure, not
on the order of insertion for dynamically inserted parts.

[1] See https://www.w3.org/2021/06/17-openui-irc

Bug: 1209219,1191121
Change-Id: Ia68c782b7972317b8e6abdcfff5e52ece34a3ed0