URLPattern: Implement compare() method.
This CL adds a prototype URLPattern.compare() to provide a natural
ordering to URLPattern objects.  This was based on feedback from routing
framework authors and there is some discussion in:

https://github.com/WICG/urlpattern/issues/61

The general algorithm is to compare the two URLPattern objects component
by component in order from hash through protocol.  Each component
pattern is then compared Part by Part.  The PartType, Modifier, and
text contents are then used to determine which Part is "more specific".

Bug: 1232795
Change-Id: I8474cd7d7689e657c9c74c552ad630cdcdd86c95