Fix Canvas 2d attribute spelling
TextRendering properties are saved as "Auto", "GeometricPrecision",
"OptimizeSpeed", etc, which is wrong and it is supposed to start with
lower case letters instead, i.e., "auto", "geometricPrecision",
"optimizeSpeed", etc. This similar mistake were made for FontKerning,
FontVariant and FontStretch.

This Cl fixes the spelling issues for these Canvas 2D attributes.

Bug: 1230242

Change-Id: I8e5e4cd60fcf75ec5de984a880c5b5b9426a75e6